Chore: Move PointManiger to PointService
