+++ src/main/java/com/challenge/devchall/point/schedule/Schedule.java
@@ -1,0 +1,4 @@
+package com.challenge.devchall.point.schedule;
+
+public class Schedule {
+}
